@@ -5,11 +5,17 @@ description = ""
 authors = ["Andrew Thwaites <acgt2@cam.ac.uk>"]
 
 [tool.poetry.dependencies]
-python = "~3.8.9"
+python = "~3.11"
+#
 colorama = "^0.4.5"
-PyYAML = "^6.0"
+llvmlite = "^0.41.0"
+mat73 = "^0.62"
 matplotlib = "^3.7.1"
+pandas = "^2.1.1"
+PyYAML = "^6.0"
 seaborn = "^0.12.2"
+sparse = "^0.14.0"
+xarray = "^2023.9.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
